Separated all proxy tests on its own file
